refactor: 불필요 인터페이스 제거
